requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Werkzeug==0.9.6
argparse==1.2.1
ashes==0.7.3
flask-mongoengine==0.7.1
flufl.enum==4.0.1
gunicorn==19.1.1
httplib2==0.9
mongoengine==0.8.7
pymongo==2.7.2
python-dateutil==2.4.0
python-loaders==0.2.3
requests==2.6.0
wsgiref==0.1.2
requests-facebook=0.2.0